requirements-ai-enhanced.txt
# Enhanced AI requirements for transformer-based models

torch>=1.13.0
transformers>=4.30.0
sentencepiece>=0.1.99
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
joblib>=1.3.0

# Additional libraries for enhanced reasoning
networkx>=3.1

# For data processing and generation
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# For handling various data formats
PyYAML>=6.0